Added a flag to test a certain behaviour
* During an `onUpdate(FrameTime frameTime)` call in `BaseArFragment`
Sceneform is asking ARCore to hand out an AugmentedImageDatabase, even
if it is null. I added for testing purposes a new flag to prevent that.
The default value is `true` to not bother the normal SDK-User with it.
So why did I do this? Well, I found in a Stacktrace the call to
`getAugmentedImageDatabase()`. And on the other hand, why do such a call
if no AugmentedImageDatabase is set.